api/requirements.txt
hiredis==2.2.2
redis==4.5.4
redis-om==0.1.2
types-redis==4.5.4.1
langchain==0.0.146
slack-bolt==1.17.2
openai==0.27.4
tiktoken==0.3.3
fastapi~=0.95.1
uvicorn==0.21.1
qdrant-client==1.1.6
prisma==0.8.2
